% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/application.R
\docType{class}
\name{Application-class}
\alias{Application}
\alias{Application-class}
\title{Generic abstract class for an application. Provides basic functions
 like logging, working with environment and error handling}
\description{
Generic abstract class for an application. Provides basic functions
 like logging, working with environment and error handling
}

