\name{keboola.r.application-package}
\alias{keboola.r.application-package}
\alias{keboola.r.application}
\docType{package}
\title{
Keboola R Application
}
\description{
Base application class, this package serves only as an ancestor for other applications, it does nothing.
}
\details{
\tabular{ll}{
Package: \tab keboola.r.application\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-04\cr
License: \tab MIT\cr
}
}
\author{
Maintainer: Ondrej Popelka <ondrej.popelka@keboola.com>
}
\keyword{ programming }
\examples{
# Subclass the application 
}
